Share PromQL set-operator lowering helpers
